GODRIVER-2724 Move cmd and benchmark to internal
