+++ .pre-commit-config.yaml
@@ -41,7 +41,7 @@ repos:
   hooks:
   - id: codespell
     args: ["-L", "te,fo,fle,alo,nin,compres,wil,collone,asess,sav,ot,wll,dne,nulll,hellow"]
-    exclude: ^(vendor/|benchmark/operation_test.go)
+    exclude: ^(vendor/|internal/benchmark/operation_test.go)
     exclude_types: [json,yaml,pem]
 
 - repo: https://github.com/golangci/golangci-lint
