+++ Makefile
@@ -112,7 +112,7 @@ build-faas-awslambda:
 ### Evergreen specific targets. ###
 .PHONY: build-aws-ecs-test
 build-aws-ecs-test:
-	go build $(BUILD_TAGS) ./cmd/testaws/main.go
+	go build $(BUILD_TAGS) ./internal/cmd/testaws/main.go
 
 .PHONY: evg-test
 evg-test:
@@ -125,7 +125,7 @@ evg-test-atlas-data-lake:
 
 .PHONY: evg-test-enterprise-auth
 evg-test-enterprise-auth:
-	go run -tags gssapi ./cmd/testentauth/main.go
+	go run -tags gssapi ./internal/cmd/testentauth/main.go
 
 .PHONY: evg-test-kmip
 evg-test-kmip:
@@ -191,7 +191,7 @@ evg-test-versioned-api:
 
 .PHONY: build-kms-test
 build-kms-test:
-	go build $(BUILD_TAGS) ./cmd/testkms
+	go build $(BUILD_TAGS) ./internal/cmd/testkms
 
 ### Benchmark specific targets and support. ###
 .PHONY: benchmark
@@ -200,7 +200,7 @@ benchmark:perf
 
 .PHONY: driver-benchmark
 driver-benchmark:perf
-	@go run cmd/godriver-benchmark/main.go | tee perf.suite
+	@go run internal/cmd/godriver-benchmark/main.go | tee perf.suite
 
 perf:driver-test-data.tar.gz
 	tar -zxf $< $(if $(eq $(UNAME_S),Darwin),-s , --transform=s)/testdata/perf/
