+++ internal/benchmark/bson_struct.go
