+++ internal/benchmark/multi.go
