+++ internal/benchmark/single.go
