+++ internal/cmd/testaws/main.go
