+++ internal/cmd/testentauth/main.go
